Arrumando as classes em arquivos
