+++ src/include/cliente.py
@@ -1,0 +1,5 @@
+from pessoa import Pessoa
+
+class Cliente(Pessoa):
+    def __init__(self):
+        super().__init__()
