--- src/include/notaFiscal.py
@@ -1,7 +1,0 @@
-class NotaFiscal:
-    def __init__(self, veiculo, valorPago, cnpj, nomeEmpresa, identificacaoComprador):
-        self._veiculo = veiculo
-        self._valorPago = valorPago
-        self._cnpj = cnpj
-        self._nomeEmpresa = nomeEmpresa
-        self._identificacaoComprador = identificacaoComprador
